Projeto utilizado no curso de MVC com PHP: entenda o padrão Model-View-Controller
