--- php-mvc-aula8-final/.gitignore
@@ -1,3 +1,0 @@
-vendor/
-.idea/
-db.sqlite
