--- php-mvc-aula8-final/cli-config.php
@@ -1,7 +1,0 @@
-<?php
-
-require_once __DIR__ . '/vendor/autoload.php';
-
-return \Doctrine\ORM\Tools\Console\ConsoleRunner::createHelperSet(
-    (new \Alura\Cursos\Infra\EntitymanagerCreator())->getEntityManager()
-);
