--- php-mvc-aula8-final/config/routes.php
@@ -1,26 +1,0 @@
-<?php
-
-use Alura\Cursos\Controller\{CursosEmJson,
-    CursosEmXml,
-    Deslogar,
-    Exclusao,
-    FormularioEdicao,
-    FormularioInsercao,
-    FormularioLogin,
-    ListarCursos,
-    Persistencia,
-    RealizarLogin};
-
-return [
-    '/listar-cursos' => ListarCursos::class,
-    '/novo-curso' => FormularioInsercao::class,
-    '/salvar-curso' => Persistencia::class,
-    '/excluir-curso' => Exclusao::class,
-    '/alterar-curso' => FormularioEdicao::class,
-    '/login' => FormularioLogin::class,
-    '/realiza-login' => RealizarLogin::class,
-    '/logout' => Deslogar::class,
-    '/buscarCursosEmJson' => CursosEmJson::class,
-    '/buscarCursosEmXml' => CursosEmXml::class
-];
-
