--- php-mvc-aula8-final/src/Controller/Deslogar.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Alura\Cursos\Controller;
-
-use Nyholm\Psr7\Response;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Server\RequestHandlerInterface;
-
-class Deslogar implements RequestHandlerInterface
-{
-    public function handle(ServerRequestInterface $request): ResponseInterface
-    {
-        session_destroy();
-        return new Response(302, ['Location' => '/login']);
-    }
-}
