--- php-mvc-aula8-final/src/Helper/FlashMessageTrait.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Alura\Cursos\Helper;
-
-trait FlashMessageTrait
-{
-    public function defineMensagem(string $tipo, string $mensagem): void
-    {
-        $_SESSION['mensagem'] = $mensagem;
-        $_SESSION['tipo_mensagem'] = $tipo;
-    }
-}
