--- php-mvc-aula8-final/src/Helper/RenderizadorDeHtmlTrait.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Alura\Cursos\Helper;
-
-trait RenderizadorDeHtmlTrait
-{
-    public function renderizaHtml(string $caminhoTemplate, array $dados): string
-    {
-        extract($dados);
-        ob_start();
-        require __DIR__ . '/../../view/' . $caminhoTemplate;
-        $html = ob_get_clean();
-
-        return $html;
-    }
-}
