--- php-mvc-aula8-final/src/Infra/EntityManagerCreator.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace Alura\Cursos\Infra;
-
-use Doctrine\ORM\EntityManager;
-use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\ORM\Tools\Setup;
-
-class EntityManagerCreator
-{
-    public function getEntityManager(): EntityManagerInterface
-    {
-        $paths = [__DIR__ . '/../Entity'];
-        $isDevMode = false;
-
-        $dbParams = array(
-            'driver' => 'pdo_sqlite',
-            'path' => __DIR__ . '/../../db.sqlite'
-        );
-
-        $config = Setup::createAnnotationMetadataConfiguration(
-            $paths,
-            $isDevMode
-        );
-        return EntityManager::create($dbParams, $config);
-    }
-}
