--- php-mvc-aula8-final/view/fim-html.php
@@ -1,3 +1,0 @@
-</div>
-</body>
-</html>
