+++ php/mvc-com-php/.gitignore
@@ -1,0 +1,3 @@
+vendor/
+.idea/
+db.sqlite
