+++ php/mvc-com-php/README.md
@@ -1,0 +1,16 @@
+# MVC com PHP: entenda o padrão Model-View-Controller
+
+Categoria | Carga horária | Atividades | Finalizado | Certificado |
+:-:|:-:|:-:|:-:|:-:|
+Programação | 14 horas | 92 | 09/04/2022 | [Certificado Normal](https://cursos.alura.com.br/certificate/22d7d603-7d92-4288-b2e7-e3b66241c086) / [Certificado Formal](https://cursos.alura.com.br/user/rodineicosta/course/php-model-view-controller/formalCertificate)
+
+## Aulas
+
+- Uma entrada para as requisições
+- Entendendo Modelo View Controller
+- HTTP, Formulários e Validação
+- Completando o cadastro
+- Autenticando usuários
+- Trabalhando com Sessão
+- PSRs e Boas Práticas
+- Introdução a WebServices
