+++ php/mvc-com-php/cli-config.php
@@ -1,0 +1,7 @@
+<?php
+
+require_once __DIR__ . '/vendor/autoload.php';
+
+return \Doctrine\ORM\Tools\Console\ConsoleRunner::createHelperSet(
+    (new \Alura\Cursos\Infra\EntitymanagerCreator())->getEntityManager()
+);
