+++ php/mvc-com-php/config/dependencies.php
@@ -1,0 +1,12 @@
+<?php
+
+$builder = new DI\ContainerBuilder();
+$builder->addDefinitions([
+    \Doctrine\ORM\EntityManagerInterface::class => function () {
+        return (new \Alura\Cursos\Infra\EntityManagerCreator())
+            ->getEntityManager();
+    }
+]);
+$container = $builder->build();
+
+return $container;
