+++ php/mvc-com-php/src/Controller/CursosEmJson.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Alura\Cursos\Controller;
+
+use Alura\Cursos\Entity\Curso;
+use Doctrine\ORM\EntityManagerInterface;
+use Nyholm\Psr7\Response;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+
+class CursosEmJson implements RequestHandlerInterface
+{
+    /**
+     * @var \Doctrine\Common\Persistence\ObjectRepository
+     */
+    private $repositorioDeCursos;
+
+    public function __construct(EntityManagerInterface $entityManager)
+    {
+        $this->repositorioDeCursos = $entityManager
+            ->getRepository(Curso::class);
+    }
+
+    public function handle(ServerRequestInterface $request): ResponseInterface
+    {
+        $cursos = $this->repositorioDeCursos->findAll();
+        return new Response(
+            200,
+            ['Content-Type' => 'application/json'],
+            json_encode($cursos)
+        );
+    }
+}
