+++ php/mvc-com-php/src/Helper/FlashMessageTrait.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Alura\Cursos\Helper;
+
+trait FlashMessageTrait
+{
+    public function defineMensagem(string $tipo, string $mensagem): void
+    {
+        $_SESSION['mensagem'] = $mensagem;
+        $_SESSION['tipo_mensagem'] = $tipo;
+    }
+}
