+++ php/mvc-com-php/view/inicio-html.php
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <title>Document</title>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+</head>
+<body>
+<?php if (isset($_SESSION['logado'])): ?>
+<nav class="navbar navbar-dark bg-dark mb-2">
+    <a class="navbar-brand" href="/listar-cursos">Home</a>
+
+    <ul class="navbar-nav ml-auto">
+        <li class="nav-item active">
+            <a class="nav-link" href="/logout">Sair</a>
+        </li>
+    </ul>
+</nav>
+<?php endif; ?>
+
+<div class="container">
+    <div class="jumbotron">
+        <h1><?= $titulo; ?></h1>
+    </div>
+
+    <?php if (isset($_SESSION['mensagem'])): ?>
+    <div class="alert alert-<?= $_SESSION['tipo_mensagem']; ?>">
+        <?= $_SESSION['mensagem']; ?>
+    </div>
+    <?php
+        unset($_SESSION['mensagem']);
+        unset($_SESSION['tipo_mensagem']);
+    endif;
+    ?>
